Completed DebuggingAPI
This code utilizes the accelerometer and the LED located on the Flight controller to indicate when the drone is moving along the Z axis (if the drone is going upwards/downwards). The green LED will turn on and the red LED will remain off, indicating that the drone is moving towards the positive Z axis (upwards). When the drone is moving towards the negative Z axis (downwards), the Red LED will light up and the green LED will turn off. The LED will continue this process until drone has been shut off. 
This projects allows us to get practice and understanding of how the sensors and LED can be utilized.