Description of project
This project was designed to allow us to practice our coding skills and working with the accelerometer. The purpose of the project is to arm the drone when it experiences free fall to keep it from crashing. This work by having the accelerometer identify when the drone has a net acceleration of zero (in this case, a small enough net acceleration) to know when to arm the drone. This means that when the drone is tossed up, it doesn't turn on until it reaches its starts to descent.