Description of project
This project was designed to allow us to practice our coding skills and work with the accelerometer. The purpose of the project is to arm the drone and prevent it from falling to the ground when it has a net acceleration of zero (in our case, an acceleration value less than two). This mean that when the drone is tossed upward, the drone will arm just as it’s about to descent. Another condition implemented to our project is to prevent the drone from arming once it has crashed to further prevent damaging the drone.