Description of project
This project allows us to expand our knowledge on how barometers work. In this project, the drone is to ascend when it has detected a change in pressure in the atmosphere. The program obtain an initial reading of the current pressure and compares that to the current pressure. If the pressure has been changed and has a difference of a value of our choosing, the drone will ascend to an altitude we have selected. On the IDE monitor, we can observe the initial and changed pressure while also see a graph of the pressure difference determined by the barometer.