Description of project
This project will consist of working with the barometer but it will be used to determine the temperature of its surrounding area. The LED on the drone and the IDE monitor will also be used. The LED will change color whenever the drone has experienced a temperature change. The IDE monitor will display a value representing the initial temperature and will change depending on the temperature.