Description of project
The following project allows the drone to flip to its starting position when it detects that the drone is upside down. We will be able to fly the drone normally once it has flipped right side up. The program allows two of its propellers to help correct the orientation of the drone. For this project, we will have to connect the propellers to the M1 – M4 motors drivers since these are bidirectional. We will assign two propellers from the same side to flip the drone over so that the drone is able to continue flying. 
What till determine when the drone is upside down is the angle of roll of the drone. The roll of the drone is between -1800 and 1800 decidegrees, when it is flipped upside down, it is between -1800 to 0 decidegrees. This will allow the drone to trigger the section of the code to flip it over, to be between 0 to 1800 decidegrees, which is the starting position for the drone.